+++ CHANGELOG.md
@@ -15,3 +15,8 @@
 ## Version 2.2 - July 31, 2013
 * Enhancement: Bootstrap 3.0 support ([PR #7](https://github.com/alefeuvre/foundation-grid-displayer/pull/7) thanks @thorsten)
 * Bugfix: Outside borders for gutterless grids (F3+, B3)
+
+## Version 3.0 - February 8, 2014
+* Enhancement: Showing "gutters" (i.e. column padding) for gutterless grid systems—Bootstrap 3, Foundation 4 & 5 ([PR #9](https://github.com/alefeuvre/foundation-grid-displayer/pull/9) thanks @JayVerb)
+* Enhancement: Changeable gutters' width
+* Enhancement: Foundation 5.0 support